requirements.txt
paddleocr>=2.6.0
pyautogui>=0.9.54
pillow>=10.0.0
selenium>=4.9.0
paddlepaddle>=2.5.0
dashscope>=1.10.0
beautifulsoup4>=4.12.0
pytest>=7.3.1
pytest-html>=3.2.0
ddt>=1.6.0
requests>=2.31.0 